Move TXO tab CSS so it can be shared easily
